+++ core/src/de/cubeisland/games/component/PlayerInput.java
@@ -8,7 +8,7 @@ import de.cubeisland.games.entity.Entity;
 public class PlayerInput extends Component<Entity> {
 
     @Override
-    public void update(int delta) {
+    public void update(float delta) {
         
     }
 }
